change type from unsigned int to size_t

+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: nde-sant <nde-sant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/15 11:31:58 by nde-sant          #+#    #+#             */
-/*   Updated: 2025/07/15 14:00:27 by nde-sant         ###   ########.fr       */
+/*   Updated: 2025/07/16 10:32:43 by nde-sant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *ptr, int c, unsigned int n)
+#include <stddef.h>
+
+void	*ft_memset(void *ptr, int c, size_t n)
 {
 	unsigned char	*dest;
 	unsigned int	i;
